Youtube_Transcript_Summary/src/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
yt-dlp
openai==1.6.1
pydantic
httpx